Add support for checkpoints
As loading quotes from wiki dumps are quite time-exhausting task,
it's very bad situation, when it fails in the middle or just before
the end of the wiki dump. So, this commit adds checkpoints, that are
created in case of an error. These checkpoints are automatically
used next time the application is started again after failure.

To enable this feature, mechanism of detecting whether the wiki dump
should be loaded or not is changed a bit. It doesn't rely on number
of quotes in database (0 for loading), but on *.done files. If there
is a *.done file for the wikidump, it won't be loaded.